fix: migration issues on laravel 11

+++ app/laravel/database/migrations/2023_02_07_130038_create_sensor_events_table.php
@@ -34,7 +34,7 @@ return new class extends Migration
             // 信号
             $table->boolean('signal')->comment(__('yokakit.signal'));
             // 値
-            $table->float('value', 8, 3);
+            $table->decimal('value', 8, 3);
             // タイムスタンプ
             $table->timestamp('at', 3)->index()->default(DB::raw('CURRENT_TIMESTAMP'));
             // 外部キー
